Delegate slf4j in HostClassLoader to system class loader (#11670)
* Delegate slf4j in HostClassLoader to system class loader

That way in Enso and Java code any usage of slf4j will use the same
configuration as the rest of the process that started it.
Tricky to test, will post an example in GH PR.

* Semi-working test

* Make sure memory appender starts clean

Appender is re-used among many tests, collecting all logs. This is
problematic when we want to test the contents of logs.