Do not report exceptions on long running Excel reads (#11916)
* Do not report exceptions on long running Excel reads

This change introduces two modifications:
- `ClosedByInterruptException` is wrapped in `InterruptedException`
  instead of `RuntimeException`
- when instrumentation encounters `InterruptedException` it bails early

Having `ClosedByInterruptException` wrapped in `RuntimeException` meant
that it is being reported as a regular `HostException` in the engine and
to the user. Instead it should be treated specially since we know that
it is caused by cancelling a long-running job. Since it is a statically
checked exception it has to be declared and the information has to be
propagated through various lambda constructs (thanks Java!).

The above change alone meant that an error is not reported for
`Data.read` nodes but any values dependent on it would still report
`No_Such_Method` error when the exception is reported as a value. Hence
the early bail out mechanism.

* Send `PendingInterrupted` on interrupt

The information could be used in GUI to indicate pending execution that
will take tad longer.

* Prettify

* Test `PendingInterrupted` payload

* Add `wasInterrupted` flag to `Pending`

Reduce `PendingInterrupted` to a flag in `Pending`

* fmt