Cleanup jline dependencies (#12020)
A wild attempt at fixing MacOS signing process that prevents nightlies from working.
Also cleaned up dependencies for wrapper projects - `scala-compiler` was unnecessary and was introducing jline requirement.
Related to #11957.

# Important Notes
Not sure if this will fix the nightly as I'm still unable to confirm the real issue but this doesn't hurt.